requirements.txt
colorama==0.4.4
dnspython==2.2.0
ipwhois==1.2.0
requests==2.26.0
